opsplitsing dialog_window + debuggen filter die kapot was
