+++ Domain/template_domain.py
@@ -17,10 +17,9 @@ class TemplateDomain:
                 return False
         return True
 
-    # TODO change selected_classes to selected_classes_dir
     # TODO: geef model_directory mee
     @staticmethod
-    def create_template(subset_path, document_path, selected_classes: List, generate_choice_list: bool,
+    def create_template(subset_path, document_path, selected_classes_dir: List, generate_choice_list: bool,
                         add_geo_artefact: bool, add_attribute_info: bool, highlight_deprecated_attributes: bool,
                         amount_of_examples: int):
         try:
@@ -28,7 +27,7 @@ class TemplateDomain:
             template_creator = SubsetTemplateCreator()
             template_creator.generate_template_from_subset(path_to_subset=subset_path,
                                                            path_to_template_file_and_extension=document_path,
-                                                           list_of_otl_objectUri=selected_classes,
+                                                           list_of_otl_objectUri=selected_classes_dir,
                                                            generate_choice_list=generate_choice_list,
                                                            add_geo_artefact=add_geo_artefact,
                                                            add_attribute_info=add_attribute_info,
